requirements.txt
scikit-learn
nltk
openai
streamlit
python-dotenv
